+++ R/teal_data-get_code.R
@@ -1,6 +1,6 @@
 #' Get code from `teal_data` object
 #'
-#' Retrieve code from `teal_data` object.
+#' #' `r lifecycle::badge("deprecated")` Retrieve code from `teal_data` object.
 #'
 #' Retrieve code stored in `@code`, which (in principle) can be used to recreate all objects found in `@env`.
 #' Use `datanames` to limit the code to one or more of the datasets enumerated in `@datanames`.
@@ -104,11 +104,14 @@
 #'
 #' @export
 setMethod("get_code", signature = "teal_data", definition = function(object, deparse = TRUE, datanames = NULL, ...) {
+
   checkmate::assert_character(datanames, min.len = 1L, null.ok = TRUE)
   checkmate::assert_flag(deparse)
 
   code <- if (!is.null(datanames)) {
+    warning("get_code(datanames) was deprecated in teal.data 0.6.1, use get_code(names) instead.")
     get_code_dependency(object@code, datanames, ...)
+
   } else {
     object@code
   }